third_party/blink/web_tests/fast/animation/scroll-animations/scroll-timeline-snapshotting.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>ScrollTimeline snapshotting</title>
<link rel="help" href="https://wicg.github.io/scroll-animations/#avoiding-cycles">
<script src='../../../resources/testharness.js'></script>
<script src='../../../resources/testharnessreport.js'></script>
<script src='../../../external/wpt/web-animations/testcommon.js'></script>
<script src='../../../external/wpt/scroll-animations/testcommon.js'></script>
<style>
  body {
    height: 800px;
    width: 800px;
  }
</style>
<div id="log"></div>

<script>
'use strict';

promise_test(async t => {
  internals.settings.setScrollAnimatorEnabled(true);

  const scroller = document.scrollingElement;
  const maxScroll = scroller.scrollHeight - scroller.clientHeight;
  const timeline = new ScrollTimeline({scrollSource: scroller});

  eventSender.keyDown("ArrowDown");
  // Wait for the scroll to change.
  const startScroll = scroller.scrollTop;
  do {
    await waitForNextFrame();
  } while(scroller.scrollTop == startScroll);

  assert_percents_equal(
      timeline.currentTime,
      (scroller.scrollTop / maxScroll) * 100,
      'Scroll timeline current time corresponds to the scroll position.');
}, 'Scroll timeline current time is updated after handling user-driven scroll.');
</script>
